ml/pyproject.toml
[project]
name = "cortex-ml"
version = "0.1.0"
description = "Cortex ML service — embeddings, reranking, and LLM orchestration"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "protobuf>=5.26.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
